requirements.txt
appnope==0.1.0
APScheduler==3.6.3
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.2.1
django-filter==1.1.0
django-registration==3.1
djangorestframework==3.7.3
dnspython==1.16.0
entrypoints==0.3
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycryptodome==3.9.7
Pygments==2.6.1
pylint==2.4.4
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.44.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
uWSGI==2.0.18
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yagmail==0.11.224
zipp==3.1.0
